🚸 Add configuration files to precompile headers
