+++ src/micras_controller.cpp
@@ -7,7 +7,6 @@
  */
 
 #include "micras/micras_controller.hpp"
-#include "target.hpp"
 
 namespace micras {
 MicrasController::MicrasController() :
